requirements.txt
torch>=1.12.0
torchvision>=0.13.0
einops>=0.6.0
numpy>=1.21.0
scipy>=1.7.0
pyyaml>=6.0
tqdm>=4.64.0
tensorboard>=2.10.0
pillow>=9.0.0
